backend/pyproject.toml
[project]
name = "smart-payment-checkout"
version = "0.1.0"
description = "Smart Payment Checkout System with AI Risk Assessment"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "openai>=1.0.0",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
